fix: プレミアム公開動画に対応するため配信中の視聴者数をStringで取得 YouTubeDataApiではプレミアム公開中の視聴者数は取得できない

+++ app/src/main/java/com/cidra/hologram/api/FirebaseService.kt
@@ -30,7 +30,7 @@ object FirebaseService {
                         it.child("title").value.toString(),
                         it.child("thumbnailUrl").value.toString(),
                         it.child("startTime").value.toString(),
-                        it.child("currentViewers").value.toString().toInt(),
+                        it.child("currentViewers").value.toString(),
                         it.child("channelName").value.toString(),
                         it.child("channelIconUrl").value.toString()
                     )
